feat: creación clase FormularioHelper para mayor modularidad y legibilidad
